refactor(types): balance mechanism lives with the data type
hydrate_track / balance_track_for_dp move to unirl.types.rollout_resp (next
to _hydrate_tensor_meta) — data reshaping lives on the data layer, matching
slice/select/concat/chunk on Batch. BaseTrainer keeps a thin opt-in policy
wrapper (balance_dp_batch flag, default False; ARTrainer is the enabler);
diffusion/PE/unified workloads are uniform-length and stay untouched.
